+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Set the default `send_attempts` to `0` (this disables retries) and deprecate the option. If you require retries you can increase the `send_attempts` option to the desired value. (#1312)
 - Bump minimum version of `guzzlehttp/psr7` package to avoid [`CVE-2022-24775`](https://github.com/guzzle/psr7/security/advisories/GHSA-q7rv-6hp3-vh96) (#1305)
+- Add `http_connect_timeout` and `http_timeout` client options (#1282)
 
 ## 3.4.0 (2022-03-14)